pyproject.toml
[project]
name = "minerreview"
version = "0.1.0"
description = "Training-free, conference-aware multi-agent peer review system"
authors = [
    { name = "peelsannaw", email = "2736522552@qq.com" }
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openreview-py>=1.32.0",
    "sentence-transformers>=2.6.1",
    "faiss-cpu>=1.8.0",
    "pydantic>=2.7.0",
    "pyyaml>=6.0.1",
    "scikit-learn>=1.4.0",
    "numpy>=1.26.0",
    "tqdm>=4.66.0",
    "requests>=2.31.0",
    "joblib>=1.3.0",
    "pandas>=2.2.0",
    "pyarrow>=15.0.0",
    "openai>=1.40.0",
    "pymilvus>=2.6.0"
]

[project.scripts]
peerreviewer = "main:main"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
packages = [{ include = "*", from = "src" }]
